+++ CHANGELOG.md
@@ -1,3 +1,21 @@
+# <a href='https://github.com/mrjackwills/flightbox_backend/releases/tag/v0.1.0'>v0.1.0</a>
+### 2023-02-02
+
+### Chores
++ dev container install typos, [c55a9d3f](https://github.com/mrjackwills/flightbox_backend/commit/c55a9d3f818af63d69dd5e1f3038809f02ac7a67)
++ github workflow on main semver only, [6dacf644](https://github.com/mrjackwills/flightbox_backend/commit/6dacf644dfd4d699d96e19697b19d81b0e37e29c)
+
+### Features
++ openssl dependency removed, [346706a9](https://github.com/mrjackwills/flightbox_backend/commit/346706a91518f9af05661328dd629df7e5d4f219)
++ create_release check for typos, [3a419dcf](https://github.com/mrjackwills/flightbox_backend/commit/3a419dcf227cc11c8268ad44b92ecbdb1a18da63)
++ use scratch docker container, [ab80ffee](https://github.com/mrjackwills/flightbox_backend/commit/ab80ffee5e813f6ad649c6d519c22b4a6dc9c7c4)
+
+### Refactors
++ tracing, [80a90454](https://github.com/mrjackwills/flightbox_backend/commit/80a90454c7d71efdb8e8f043c1619b4cd6a325c3)
+
+### Tests
++ incoming message tests, [2e0645ef](https://github.com/mrjackwills/flightbox_backend/commit/2e0645ef8fa2d13e3c04084cdccfb449250ceaff)
+
 # <a href='https://github.com/mrjackwills/flightbox_backend/releases/tag/v0.0.2'>v0.0.2</a>
 ### 2023-01-24